Move types and methods to shcd package
